Restrukturer litt for å skille mellom inntektsmeldinger som leses og inntektsmeldinger som blir sendt inn
