+++ src/main/kotlin/no/nav/helsearbeidsgiver/innsending/InnsendingService.kt
@@ -2,7 +2,7 @@ package no.nav.helsearbeidsgiver.innsending
 
 import no.nav.helsearbeidsgiver.domene.inntektsmelding.v1.skjema.SkjemaInntektsmelding
 import no.nav.helsearbeidsgiver.innsending.Innsending.toJson
-import no.nav.helsearbeidsgiver.kafka.inntektsmelding.InnsendingProducer
+import no.nav.helsearbeidsgiver.kafka.innsending.InnsendingProducer
 import no.nav.helsearbeidsgiver.utils.json.serializer.LocalDateTimeSerializer
 import no.nav.helsearbeidsgiver.utils.json.serializer.UuidSerializer
 import no.nav.helsearbeidsgiver.utils.json.toJson
