+++ src/main/resources/application.conf
@@ -6,6 +6,12 @@ kafkaConsumer {
   forespoersel {
     topic = helsearbeidsgiver.pri
   }
+
+}
+kafkaProducer {
+  innsending {
+    topic = helsearbeidsgiver.api-innsending
+  }
 }
 database {
   username = "lpsapi"
